UDP send socket

+++ src/libc/micros/sys/micros_socket.h
@@ -34,12 +34,12 @@ int socket(int domain, int type, int protocol);
 
 int bind(int s, struct sockaddr *my_addr, socklen_t addrlen);
 
-ssize_t recv(int s, void *buf, size_t len, int flags);
+uint32_t recv(int s, void *buf, size_t len, int flags);
 
-ssize_t recvfrom(int s, void *buf, size_t len, int flags,
+uint32_t recvfrom(int s, void *buf, size_t len, int flags,
                  struct sockaddr *from, socklen_t *fromlen);
 
-ssize_t sendto(int s, const void *buf, size_t len,
+uint32_t sendto(int s, const void *buf, size_t len,
                int flags, const struct sockaddr *to,
                socklen_t tolen);
